shop update

+++ database/models.py
@@ -1,5 +1,4 @@
-
-from sqlalchemy import Boolean, Column, Integer, String, DateTime
+from sqlalchemy import Boolean, Column, Integer, BigInteger, String, DateTime
 from database.database import Base
 
 class User(Base):
@@ -8,11 +7,11 @@ class User(Base):
     id = Column(Integer, primary_key=True, index=True)
     username = Column(String, unique=True, index=True)
     password = Column(String)
-    balance = Column(Integer, default=0)
+    balance = Column(BigInteger, default=0)
     is_admin = Column(Boolean, default=False)
     id_banned = Column(Boolean, default=False)
     show_in_leaderboard = Column(Boolean, default=True)
-    total_gold_earned = Column(Integer, default=0)
+    total_gold_earned = Column(BigInteger, default=0)
     last_dayle = Column(DateTime)
     buzz_coins = Column(Integer, default=0)
     
@@ -22,10 +21,9 @@ class game_history(Base):
     id = Column(Integer, primary_key=True, index=True)
     username = Column(String)
     game = Column(String)
-    balance = Column(Integer)
+    balance = Column(BigInteger)
     time = Column(String)
     win = Column(Boolean)
-    
     
 class Settings(Base):
     __tablename__="Settings"
@@ -36,7 +34,6 @@ class Settings(Base):
     murmel_duration = Column(Integer, default=5)    
     custom_input = Column(Boolean, default=False)
     
-
 class coin_codes(Base):
     __tablename__="coin_codes"
